Add files via upload
SearchEngine.java is the main class.
action.txt contains the action to be done. Every new action should be added to it.
A folder with the name 'webpages' is to be created in the same folder as the java files and the files used for search engine should be copied there.